Revert "Set feeIncreaseFactor type to float"
This reverts commit 94b9586a192d16ac06e134e7b43a49c57ac68ad0.

+++ node-it/src/test/scala/com/wavesplatform/it/api/AsyncHttpApi.scala
@@ -469,7 +469,7 @@ object AsyncHttpApi extends Assertions {
         payment: Seq[InvokeScriptTransaction.Payment] = Seq.empty,
         fee: Long = 500000,
         feeAssetId: Option[String] = None,
-        feeIncreaseFactor: Float,
+        feeIncreaseFactor: Int,
         version: TxVersion = TxVersion.V1
     ): Future[(Transaction, JsValue)] =
       signedTraceBroadcast(
@@ -496,7 +496,7 @@ object AsyncHttpApi extends Assertions {
         payment: Seq[InvokeScriptTransaction.Payment] = Seq.empty,
         fee: Long = 500000,
         feeAssetId: Option[String] = None,
-        feeIncreaseFactor: Float,
+        feeIncreaseFactor: Int,
         version: TxVersion = TxVersion.V1
     ): Future[(JsValue, JsValue)] = {
       val jsObject = InvokeScriptTransaction(
